Only try and create HTML logs if logs exist
This just cleans up an annoying error message when the tests fail
somehow, or compilation fails, and no egre.log file is created.

+++ t
@@ -17,5 +17,10 @@ fi
 CT_OPTS="-config test/test.config ${CASES} -erl_args -config rel/sys.config " \
   EGRE_LOG_PATH=$(pwd)/logs \
   make ct | tee  out
-cat log_wrappers/json_head <(sed -e 's/.*/&,/' logs/egre.log) log_wrappers/json_tail > logs/egre_mud_1_log.js
-cp log_wrappers/*.{js,css,html} logs/
+
+if [[ -e logs/egre.log ]] ; then
+  cat log_wrappers/json_head <(sed -e 's/.*/&,/' logs/egre.log) log_wrappers/json_tail > logs/egre_mud_1_log.js
+  cp log_wrappers/*.{js,css,html} logs/
+else
+  echo "No logs/egre.log file found"
+fi
